Remove YAML configuration from Withings
